requirements.txt
# FastAPI e dependências
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Análise de dados e visualização
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning
scikit-learn==1.3.2
imbalanced-learn==0.11.0
joblib==1.3.2

# Deep Learning
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0

# Utilitários
python-dotenv==1.0.0

# Adicionando pydantic
pydantic==2.5.2
